IPTE-165: Refactoring Transfer to own packages
