+++ src/main/java/de/doubleslash/keeptime/REST_API/controller/ProjectController.java
@@ -1,13 +1,13 @@
 package de.doubleslash.keeptime.REST_API.controller;
 
 import de.doubleslash.keeptime.REST_API.ProjectColorDTO;
+import de.doubleslash.keeptime.REST_API.mapper.ProjectMapper;
 import de.doubleslash.keeptime.controller.Controller;
 import de.doubleslash.keeptime.model.Model;
 import de.doubleslash.keeptime.model.Project;
 import de.doubleslash.keeptime.model.Work;
 import de.doubleslash.keeptime.model.repos.ProjectRepository;
 import de.doubleslash.keeptime.model.repos.WorkRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -17,7 +17,6 @@ import javax.validation.Valid;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 @RestController
 @RequestMapping("/projects")
