+++ src/main/java/de/doubleslash/keeptime/REST_API/mapper/ProjectMapper.java
@@ -1,8 +1,7 @@
-package de.doubleslash.keeptime.REST_API.controller;
+package de.doubleslash.keeptime.REST_API.mapper;
 import de.doubleslash.keeptime.REST_API.ProjectColorDTO;
 import de.doubleslash.keeptime.model.Project;
 import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
 import org.mapstruct.factory.Mappers;
 
 @Mapper(uses = ColorMapper.class)
