+++ src/test/java/de/doubleslash/keeptime/REST_API/controller/ProjectMapperTest.java
@@ -1,6 +1,7 @@
 package de.doubleslash.keeptime.REST_API.controller;
 
 import de.doubleslash.keeptime.REST_API.ProjectColorDTO;
+import de.doubleslash.keeptime.REST_API.mapper.ProjectMapper;
 import de.doubleslash.keeptime.model.Project;
 import javafx.scene.paint.Color;
 import org.junit.jupiter.api.Test;
