+++ src/test/java/de/doubleslash/keeptime/REST_API/controller/WorkMapperTest.java
@@ -1,16 +1,11 @@
 package de.doubleslash.keeptime.REST_API.controller;
 
+import de.doubleslash.keeptime.REST_API.DTO.WorkDTO;
+import de.doubleslash.keeptime.REST_API.mapper.WorkMapper;
 import de.doubleslash.keeptime.model.Project;
 import de.doubleslash.keeptime.model.Work;
 import javafx.scene.paint.Color;
-import net.bytebuddy.asm.Advice;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.time.LocalDateTime;
